Schema printing commands to run under the 'schema' command
